feat(crud): complete product_crud with full test coverage
- fix ProductUpdate.price type (int -> float)
- refactor update_product to accept product_id instead of Product object, matching CRUD pattern
- add remaining tests: get_products_by_category, update_product, delete_product, deduct_stock

+++ schemas/product.py
@@ -34,5 +34,5 @@ class ProductResponse(BaseModel):
 
 class ProductUpdate(BaseModel):
     name: str | None = None
-    price: int | None = None
+    price: float | None = None
     stock: int | None = None
